Refactoring project structure, onboarding added, and some widgets upgraded 🚀
