+++ lib/app/app.dart
@@ -1,6 +1,6 @@
 import 'package:dragon_store/config/foundations/theme/app_theme.dart';
 import 'package:dragon_store/config/router/app_router.dart';
-import 'package:dragon_store/ui/providers/dragon_list_provider.dart';
+import 'package:dragon_store/domain/providers/dragon_list_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -9,8 +9,8 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [ChangeNotifierProvider(create: (_) => DragonListProvider())],
+    return ChangeNotifierProvider(
+      create: (_) => DragonListProvider(),
       child: MaterialApp.router(
         routerConfig: appRouter,
         debugShowCheckedModeBanner: false,
