+++ lib/domain/entities/dragon_element_entity.dart
@@ -1,0 +1,1 @@
+enum DragonElementEntity { earth, fire, water, nature }
