+++ lib/domain/providers/detail_provider.dart
@@ -1,17 +1,17 @@
-import 'package:dragon_store/domain/entities/dragon.dart';
+import 'package:dragon_store/domain/entities/dragon_entity.dart';
 import 'package:flutter/material.dart';
 
 class DetailProvider extends ChangeNotifier {
-  late Dragon dragonSend;
+  late DragonEntity dragonSend;
   bool isReadArguments = false;
-  void setParameters(Dragon dragon) {
+  void setParameters(DragonEntity dragon) {
     if (!isReadArguments) {
       dragonSend = dragon;
       isReadArguments = true;
     }
   }
 
-  void updateParameters(Dragon dragon) {
+  void updateParameters(DragonEntity dragon) {
     dragonSend = dragon;
     notifyListeners();
   }
