+++ lib/domain/providers/form_provider.dart
@@ -1,13 +1,13 @@
-import 'package:dragon_store/domain/entities/dragon.dart';
-import 'package:dragon_store/domain/entities/dragon_element.dart';
+import 'package:dragon_store/domain/entities/dragon_entity.dart';
+import 'package:dragon_store/domain/entities/dragon_element_entity.dart';
 import 'package:flutter/material.dart';
 
 class FormProvider extends ChangeNotifier {
-  String nameDragon = '', description = '', level = '';
-  ElementDragon elementSelected = ElementDragon.earth;
+  String? nameDragon, description, level;
+  DragonElementEntity elementSelected = DragonElementEntity.earth;
   bool hasGetParameters = false;
 
-  void getParameters(Dragon? dragon) {
+  void getParameters(DragonEntity? dragon) {
     if (dragon != null && !hasGetParameters) {
       nameDragon = dragon.name;
       description = dragon.description;
@@ -17,32 +17,31 @@ class FormProvider extends ChangeNotifier {
     }
   }
 
-  Dragon? createDragon() {
-    if (nameDragon.isEmpty || description.isEmpty || level.isEmpty) return null;
-    return Dragon(
-        description: description,
-        name: nameDragon,
-        level: int.parse(level),
+  DragonEntity? createDragon() {
+    if (nameDragon?.isEmpty == true ||
+        description?.isEmpty == true ||
+        level?.isEmpty == true) return null;
+    return DragonEntity(
+        description: description!,
+        name: nameDragon!,
+        level: int.parse(level!),
         element: elementSelected);
   }
 
-  void updateElement(ElementDragon newElement) {
+  void updateElement(DragonElementEntity newElement) {
     elementSelected = newElement;
     notifyListeners();
   }
 
   void setNameDragon(String nameDragon) {
     this.nameDragon = nameDragon;
-    notifyListeners();
   }
 
   void setDescriptionDragon(String descriptionDragon) {
     description = descriptionDragon;
-    notifyListeners();
   }
 
   void setLevelDragon(String levelDragon) {
     level = levelDragon;
-    notifyListeners();
   }
 }
