+++ lib/ui/widgets/shared/element_item_view.dart
@@ -1,8 +1,9 @@
-import 'package:dragon_store/domain/entities/dragon_element.dart';
+import 'package:dragon_store/domain/entities/dragon_element_entity.dart';
+import 'package:dragon_store/domain/utils/dragon_element_utils.dart';
 import 'package:flutter/material.dart';
 
 class ElementItemView extends StatelessWidget {
-  final ElementDragon element;
+  final DragonElementEntity element;
   const ElementItemView({super.key, required this.element});
 
   @override
@@ -14,9 +15,9 @@ class ElementItemView extends StatelessWidget {
         decoration: BoxDecoration(
             borderRadius: const BorderRadius.all(Radius.circular(25)),
             color:
-                ElementDragonUtils.getColorElement(element).withOpacity(0.5)),
+                DragonElementUtils.getColorElement(element).withOpacity(0.5)),
         child: Image(
-          image: AssetImage(ElementDragonUtils.getIcon(element)),
+          image: AssetImage(DragonElementUtils.getIcon(element)),
           height: 40,
           width: 40,
         ),
