+++ pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   provider: ^6.1.2
   flutter_svg: ^2.0.10+1
   flutter_slidable: ^3.1.0
+  lottie: ^3.1.1
 
 dev_dependencies:
   flutter_test:
@@ -66,6 +67,7 @@ flutter:
   assets:
     - assets/icons/
     - assets/images/
+    - assets/animations/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
